Fixed k8s prereq

+++ scripts/proxmox-ubuntu-k8s-prereq.sh
@@ -3,11 +3,14 @@
 # This is a script to install the prerequisites for Kubernetes on Ubuntu 24.04 using kubeadm
 
 # Static IP configuration
-vm_ip=$(hostname -I)
-vm_nameserver=$(yq '.network.deco_mesh.nameserver' env.yaml)
-vm_gateway=$(yq '.network.deco_mesh.gateway' env.yaml)
+vm_ip=$(hostname -I | awk '{print $1}')
+vm_nameserver=$(yq '.networks.deco_mesh.nameserver' env.yaml)
+vm_gateway=$(yq '.networks.deco_mesh.gateway' env.yaml)
 # Create backup of the original netplan config
-cp /etc/netplan/50-cloud-init.yaml /etc/netplan/50-cloud-init.yaml.bak
+if [ ! -f /etc/netplan/50-cloud-init.yaml.bak ]; then
+  echo "Backing up existing netplan configuration..."
+  cp /etc/netplan/50-cloud-init.yaml /etc/netplan/50-cloud-init.yaml.bak
+fi
 cat >/etc/netplan/50-cloud-init.yaml <<EOL
 network:
   version: 2
@@ -34,6 +37,8 @@ echo "br_netfilter" >> /etc/modules-load.d/k8s.conf
 # Setup Kubernetes keyring (from https://kubernetes.io/docs/tasks/tools/install-kubectl-linux/#install-using-native-package-management)
 curl -fsSL https://pkgs.k8s.io/core:/stable:/v1.32/deb/Release.key | sudo gpg --dearmor -o /etc/apt/keyrings/kubernetes-apt-keyring.gpg
 chmod 644 /etc/apt/keyrings/kubernetes-apt-keyring.gpg # allow unprivileged APT programs to read this keyring
+echo 'deb [signed-by=/etc/apt/keyrings/kubernetes-apt-keyring.gpg] https://pkgs.k8s.io/core:/stable:/v1.32/deb/ /' | sudo tee /etc/apt/sources.list.d/kubernetes.list
+sudo chmod 644 /etc/apt/sources.list.d/kubernetes.list   # helps tools such as command-not-found to work correctly
 
 apt update
 apt install kubectl kubeadm kubelet
